ATOC/ATOCUnit Updates
Removed unnessary code from the atoc class, due to the condensing of the code and re-arranging of the code. Then Implemented some more tests, to ensure that the modified code is running correctly. Still need to finish implementing some of the unit tests: Projection, Rogue Detections tests. I addition, I need to run the Unit Tests to see if any of the tests are failing. After this, I will move to the graphs.