Finished Debugging Tracker Class
Finished implementing the Tracker class, and testing it will an actual simulation flight. Created a Tutorial live function that shows how the tracker class is created and used to update the current simulation step. Then it shows how well the tracker object works on the UAS flight trajectory and shows the differences between reported and the kalman prediction metrics.